+++ src/Elders.Pandora/Properties/AssemblyInfo.cs
@@ -4,11 +4,11 @@ using System.Reflection;
 [assembly: AssemblyTitleAttribute("Elders.Pandora")]
 [assembly: AssemblyDescriptionAttribute("Elders.Pandora")]
 [assembly: AssemblyProductAttribute("Elders.Pandora")]
-[assembly: AssemblyVersionAttribute("0.1.3")]
-[assembly: AssemblyInformationalVersionAttribute("0.1.3")]
-[assembly: AssemblyFileVersionAttribute("0.1.3")]
+[assembly: AssemblyVersionAttribute("0.1.4")]
+[assembly: AssemblyInformationalVersionAttribute("0.1.4")]
+[assembly: AssemblyFileVersionAttribute("0.1.4")]
 namespace System {
     internal static class AssemblyVersionInformation {
-        internal const string Version = "0.1.3";
+        internal const string Version = "0.1.4";
     }
 }