+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
+#### 0.1.4 - 14.11.2014
+* Added ApplicationConfiguration for clients to get setting value by key
+
 #### 0.1.3 - 14.11.2014
-* Add NameBuilder for files and setting key names
+* Added NameBuilder for files and setting key names
 
 #### 0.1.2 - 14.11.2014
 * Fix misspelled name in the exe file